+++ src/Gyselroth/Helper/HelperDate.php
@@ -148,11 +148,13 @@ class HelperDate implements ConstantsUnitsOfTimeInterface
      */
     public static function getZendDateByDateString(string $date): \Zend_Date
     {
-        return new \Zend_Date([
-            'year'  => \substr($date, 0, 4),
-            'month' => \substr($date, 5, 2),
-            'day'   => \substr($date, 8, 2)
-        ], self::DEFAULT_LOCALE);
+        return new \Zend_Date(
+            [
+                'year' => \substr($date, 0, 4),
+                'month' => \substr($date, 5, 2),
+                'day' => \substr($date, 8, 2)
+            ], self::DEFAULT_LOCALE
+        );
     }
 
     /**
@@ -247,9 +249,11 @@ class HelperDate implements ConstantsUnitsOfTimeInterface
      */
     public static function getDateTime($time): \DateTime
     {
-        return \is_numeric($time)
-            ? new \DateTime('@' . (int)$time)
-            : new \DateTime($time);
+        return new \DateTime(
+            \is_numeric($time)
+                ? '@' . (int)$time
+                : $time
+        );
     }
 
     /**
@@ -326,7 +330,8 @@ class HelperDate implements ConstantsUnitsOfTimeInterface
             0,
             $dateParts['month'],
             $dateParts[self::DATE_TIME_PART_DAY],
-            $dateParts[self::DATE_TIME_PART_YEAR]);
+            $dateParts[self::DATE_TIME_PART_YEAR]
+        );
     }
 
     /**
@@ -343,7 +348,8 @@ class HelperDate implements ConstantsUnitsOfTimeInterface
             59,
             $dateParts['month'],
             $dateParts[self::DATE_TIME_PART_DAY],
-            $dateParts[self::DATE_TIME_PART_YEAR]);
+            $dateParts[self::DATE_TIME_PART_YEAR]
+        );
     }
 
     /**
@@ -428,7 +434,8 @@ class HelperDate implements ConstantsUnitsOfTimeInterface
             * self::SECONDS_MIN
             + (array_key_exists('seconds', $timeParts)
                 ? $timeParts['seconds']
-                : 0);
+                : 0
+            );
     }
 
     /**
@@ -457,7 +464,8 @@ class HelperDate implements ConstantsUnitsOfTimeInterface
     {
         return self::getSumSecondsOfTimeParts(
             self::getTimeStringParts($timeStr, ':', $includeMinutes, $includeSeconds),
-            $isCurrentDate);
+            $isCurrentDate
+        );
     }
 
     /**
@@ -572,7 +580,8 @@ class HelperDate implements ConstantsUnitsOfTimeInterface
             default:
                 LoggerWrapper::warning(
                     "Detected unhandled unit $unit",
-                    [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY, LoggerWrapper::OPT_PARAMS => $unit]);
+                    [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY, LoggerWrapper::OPT_PARAMS => $unit]
+                );
 
                 return null;
         }
@@ -619,7 +628,8 @@ class HelperDate implements ConstantsUnitsOfTimeInterface
             0,
             \date('m', $dateFrom),
             \date('d', $dateFrom),
-            \date('Y', $dateFrom));
+            \date('Y', $dateFrom)
+        );
 
         $diff = $dateTo - $dateFrom;
         $days = (int)($diff / self::SECONDS_DAY);
@@ -722,7 +732,8 @@ class HelperDate implements ConstantsUnitsOfTimeInterface
             \date('Ymd', $timestamp)
           . ($appendTime
                 ? 'T' . \date('His', $timestamp)
-                : '');
+                : ''
+            );
     }
 
     /**
@@ -796,32 +807,40 @@ class HelperDate implements ConstantsUnitsOfTimeInterface
         switch ($shiftingMode) {
             case self::SHIFT_MODE_YESTERDAY:
                 $date->subDay(1);
+
                 break;
             case self::SHIFT_MODE_TODAY:
                 break;
             case self::SHIFT_MODE_TOMORROW:
                 $date->addDay(1);
+
                 break;
             case self::SHIFT_MODE_DAY_AFTER_TOMORROW:
                 $date->addDay(2);
                 $amountDaysAdded = 2;
+
                 break;
             case self::SHIFT_MODE_3_DAYS_LATER:
                 $date->addDay(3);
                 $amountDaysAdded = 3;
+
                 break;
             case self::SHIFT_MODE_4_DAYS_LATER:
                 $date->addDay(4);
                 $amountDaysAdded = 4;
+
                 break;
             case self::SHIFT_MODE_5_DAYS_LATER:
                 $date->addDay(5);
                 $amountDaysAdded = 5;
+
                 break;
             default:
                 LoggerWrapper::warning(
                     __CLASS__ . '::' . __FUNCTION__ . " - Unknown shifting mode $shiftingMode",
-                    [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY, LoggerWrapper::OPT_PARAMS => $shiftingMode]);
+                    [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY, LoggerWrapper::OPT_PARAMS => $shiftingMode]
+                );
+
                 break;
         }
 
@@ -832,6 +851,7 @@ class HelperDate implements ConstantsUnitsOfTimeInterface
 
                 if (6 === $weekdayIndex || 0 === $weekdayIndex) {
                     $hasPassedWeekend = true;
+
                     break;
                 }
             }