+++ src/Gyselroth/Helper/HelperTimerange.php
@@ -56,17 +56,22 @@ class HelperTimerange
         $dateStart = new \Zend_Date(
             HelperDate::getDateFromUnixTimestamp($dateStart),
             \Zend_Date::ISO_8601,
-            $locale);
+            $locale
+        );
 
         /** @noinspection CallableParameterUseCaseInTypeContextInspection */
         $dateEnd = new \Zend_Date(
             HelperDate::getDateFromUnixTimestamp($dateEnd),
             \Zend_Date::ISO_8601,
-            $locale);
+            $locale
+        );
 
         // @todo: "bis" needs to be translated as well if locale is set to "en"
 
-        return $dateStart->get(\Zend_Date::DATE_LONG) . ' ' . 'bis' . ' ' . $dateEnd->get(\Zend_Date::DATE_LONG);
+        return $dateStart->get(
+            \Zend_Date::DATE_LONG) . ' '
+            . 'bis' . ' '
+            . $dateEnd->get(\Zend_Date::DATE_LONG);
     }
 
     /**
@@ -149,6 +154,7 @@ class HelperTimerange
         $rangeStart = HelperDate::getSumMinutesOfTimeString($range[0]);
         $rangeEnd   = HelperDate::getSumMinutesOfTimeString($range[1]);
         $y          = 1;
+
         \imageline($image, $rangeStart, $y, $rangeEnd, $y, $black);
 
         // Draw comparison ranges
@@ -163,6 +169,7 @@ class HelperTimerange
                 $y,
                 $black
             );
+
             $y += $stepSize;
         }
 
@@ -175,6 +182,7 @@ class HelperTimerange
 
             for ($y = 1; $y <= $amountRanges; $y++) {
                 $scannedColor = \imagecolorat($image, $x, $y);
+
                 if ($scannedColor !== $white) {
                     $maxOverlapsCurrent++;
                 }
@@ -199,8 +207,9 @@ class HelperTimerange
     public static function getWeekDaysByTimeRange($startTime, $endTime): array
     {
         $weekDays = [];
-        $day      = new \Zend_Date($startTime);
-        $end      = (new \Zend_Date($endTime))->toValue();
+
+        $day = new \Zend_Date($startTime);
+        $end = (new \Zend_Date($endTime))->toValue();
 
         while ($day->toValue() <= $end) {
             $weekDays[] = $day->get(\Zend_Date::WEEKDAY_DIGIT);