+++ src/Gyselroth/Helper/HelperZip.php
@@ -152,7 +152,8 @@ class HelperZip
         if (!\file_exists($pathUnzipped)) {
             LoggerWrapper::error(
                 'Zipping failed: file does not exist:' . $pathUnzipped,
-                [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY]);
+                [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY]
+            );
 
             return false;
         }
@@ -160,7 +161,8 @@ class HelperZip
         if (\is_dir($destinationFilename)) {
             LoggerWrapper::error(
                 'Zipping failed: given destination is directory:' . $destinationFilename,
-                [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY]);
+                [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY]
+            );
 
             return false;
         }
@@ -170,7 +172,8 @@ class HelperZip
         if (!$zip->open($destinationFilename, \ZipArchive::CREATE)) {
             LoggerWrapper::error(
                 'Zipping failed: cannot open archive:' . $pathUnzipped,
-                [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY]);
+                [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY]
+            );
 
             return false;
         }
@@ -183,12 +186,14 @@ class HelperZip
             ) {
                 LoggerWrapper::warning(
                     'ZipArchive->addFile failed for file: ' . $pathUnzipped,
-                    [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY]);
+                    [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY]
+                );
             }
         } elseif (\is_dir($pathUnzipped)) {
             $files    = new \RecursiveIteratorIterator(
                 new \RecursiveDirectoryIterator($pathUnzipped),
-                \RecursiveIteratorIterator::SELF_FIRST);
+                \RecursiveIteratorIterator::SELF_FIRST
+            );
 
             $pathDots = ['.', '..'];
 
@@ -227,6 +232,7 @@ class HelperZip
                 }
             }
         }
+
         $zip->close();
 
         if ($deleteUnzipped) {
@@ -252,7 +258,8 @@ class HelperZip
         if (!\is_file($pathArchive)) {
             LoggerWrapper::error(
                 'ZIP archive doesn\'t exist: ' . $pathArchive,
-                [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY]);
+                [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY]
+            );
 
             return false;
         }
@@ -261,7 +268,8 @@ class HelperZip
         if (0 === \filesize($pathArchive)) {
             LoggerWrapper::error(
                 'Cannot unzip archive with file size of 0: ' . $pathArchive,
-                [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY]);
+                [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY]
+            );
 
             return false;
         }
@@ -275,7 +283,8 @@ class HelperZip
         if (!$zipArchive->open($pathArchive)) {
             LoggerWrapper::error(
                 'Failed opening ZIP archive: ' . $pathArchive,
-                [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY]);
+                [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY]
+            );
 
             return false;
         }
@@ -297,7 +306,8 @@ class HelperZip
 
         LoggerWrapper::error(
             'Failed extracting ZIP archive: ' . $pathArchive . ' to path: ' . $pathDestination,
-            [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY]);
+            [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY]
+        );
 
         return false;
     }
@@ -323,7 +333,8 @@ class HelperZip
 
         if (!$zip->open($pathArchive)) {
             LoggerWrapper::error('Failed opening ZIP archive: ' . $pathArchive,
-                [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY]);
+                [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY]
+            );
 
             return false;
         }
@@ -394,6 +405,7 @@ class HelperZip
         LoggerWrapper::log(
             $message,
             $logLevel,
-            [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY, 'trace' => debug_backtrace()]);
+            [LoggerWrapper::OPT_CATEGORY => self::LOG_CATEGORY, 'trace' => debug_backtrace()]
+        );
     }
 }